utilities/requirements.txt
pandas
asyncio
aiohttp
python-dotenv
BeautifulSoup4
tqdm
requests